# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Election Results API
 * OpenAPI spec version: 0.1.0
 */
export interface CreatePartyRequest {
  name: string;
  abbreviation: string;
  color: string;
  description?: string;
}

export interface HealthStatus {
  status: string;
}

export interface Election {
  id: number;
  name: string;
  date: string;
  totalSeats: number;
  /** declared | counting | pending */
  status: string;
  scope: string;
  /** @nullable */
  state?: string | null;
}

export interface ElectionSummary {
  electionId: number;
  electionName: string;
  totalRegisteredVoters: number;
  totalVotesCast: number;
  turnoutPercent: number;
  seatsTotal: number;
  seatsDeclared: number;
  /** @nullable */
  leadingParty: string | null;
  /** @nullable */
  leadingPartyColor?: string | null;
  /** @nullable */
  leadingPartySeats?: number | null;
  majorityThreshold?: number;
}

export interface PartySeatCount {
  partyId: number;
  partyName: string;
  partyAbbreviation: string;
  partyColor: string;
  seatsWon: number;
}

export interface PartyVoteShare {
  partyId: number;
  partyName: string;
  partyAbbreviation: string;
  partyColor: string;
  totalVotes: number;
  voteSharePercent: number;
}

export interface Party {
  id: number;
  name: string;
  abbreviation: string;
  color: string;
  /** @nullable */
  description?: string | null;
}

export interface ConstituencyResult {
  id: number;
  name: string;
  region: string;
  code: string;
  registeredVoters: number;
  votesCast: number;
  spoiltVotes: number;
  turnoutPercent: number;
  /** declared | counting | pending */
  status: string;
  scope: string;
  state: string;
  /** @nullable */
  winningPartyId?: number | null;
  /** @nullable */
  winningPartyName?: string | null;
  /** @nullable */
  winningPartyColor?: string | null;
  /** @nullable */
  winningPartyAbbreviation?: string | null;
  /** @nullable */
  winningCandidateName?: string | null;
  /** @nullable */
  winningVotes?: number | null;
  /** @nullable */
  margin?: number | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  gridX?: number | null;
  /** @nullable */
  gridY?: number | null;
}

export interface CandidateResult {
  id: number;
  name: string;
  partyId: number;
  partyName: string;
  partyAbbreviation: string;
  partyColor: string;
  votes: number;
  voteSharePercent: number;
  isWinner: boolean;
}

export interface ConstituencyDetail {
  id: number;
  name: string;
  region: string;
  code: string;
  registeredVoters: number;
  votesCast: number;
  spoiltVotes: number;
  turnoutPercent: number;
  status: string;
  scope: string;
  state: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  gridX?: number | null;
  /** @nullable */
  gridY?: number | null;
  candidates: CandidateResult[];
}

export type ListConstituenciesParams = {
electionId?: number;
};