# lib/api-zod/src/generated/types/constituencyDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Election Results API
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateResult } from './candidateResult';

export interface ConstituencyDetail {
  id: number;
  name: string;
  region: string;
  code: string;
  registeredVoters: number;
  votesCast: number;
  spoiltVotes: number;
  turnoutPercent: number;
  status: string;
  scope: string;
  state: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  gridX?: number | null;
  /** @nullable */
  gridY?: number | null;
  candidates: CandidateResult[];
}