# lib/api-zod/src/generated/types/constituencyResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Election Results API
 * OpenAPI spec version: 0.1.0
 */

export interface ConstituencyResult {
  id: number;
  name: string;
  region: string;
  code: string;
  registeredVoters: number;
  votesCast: number;
  spoiltVotes: number;
  turnoutPercent: number;
  /** declared | counting | pending */
  status: string;
  scope: string;
  state: string;
  /** @nullable */
  winningPartyId?: number | null;
  /** @nullable */
  winningPartyName?: string | null;
  /** @nullable */
  winningPartyColor?: string | null;
  /** @nullable */
  winningPartyAbbreviation?: string | null;
  /** @nullable */
  winningCandidateName?: string | null;
  /** @nullable */
  winningVotes?: number | null;
  /** @nullable */
  margin?: number | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  gridX?: number | null;
  /** @nullable */
  gridY?: number | null;
}