# lib/api-zod/src/generated/types/createPartyRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Election Results API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePartyRequest {
  name: string;
  abbreviation: string;
  color: string;
  description?: string;
}