# lib/api-zod/src/generated/types/election.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Election Results API
 * OpenAPI spec version: 0.1.0
 */

export interface Election {
  id: number;
  name: string;
  date: string;
  totalSeats: number;
  /** declared | counting | pending */
  status: string;
  scope: string;
  /** @nullable */
  state?: string | null;
}